[SRLT-155] Refactor: event가 application 레이어임에 따라 DTO 명 및 패키지 위치 수정
