--- src/main/java/starlight/application/aireport/event/PdfReportRequestedEvent.java
@@ -1,4 +1,0 @@
-package starlight.application.aireport.event;
-
-public record PdfReportRequestedEvent(long businessPlanId, String pdfUrl, long memberId) {
-}
