+++ src/main/java/starlight/application/aireport/event/PdfReportRequestedInput.java
@@ -1,0 +1,4 @@
+package starlight.application.aireport.event;
+
+public record PdfReportRequestedInput(long businessPlanId, String pdfUrl, long memberId) {
+}
